refactored gas calculation to run for deposits as well as send
